+++ packages/frontend/components/collection/add/AddCollectionModal.tsx
@@ -1,12 +1,5 @@
+import { Modal, ModalCloseButton, ModalHeader } from '@components/modal';
 import React from 'react';
-import {
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalCloseButton,
-  ModalBody,
-} from '@chakra-ui/react';
 import { AddCollectionForm, AddCollectionFormProps } from './AddCollectionForm';
 
 export type AddCollectionModalProps = AddCollectionFormProps & {
@@ -22,21 +15,13 @@ export const AddCollectionModal: React.FC<AddCollectionModalProps> = (
 
   return (
     <Modal
-      isOpen={isOpen}
-      onClose={onClose}
-      isCentered
-      closeOnEsc={false}
-      closeOnOverlayClick={false}
-      size={{ base: 'full', sm: 'lg' }}
+      open={isOpen}
+      onOpenChange={onClose}
+      contentTestId="addCollectionModal"
     >
-      <ModalOverlay />
-      <ModalContent data-test="addCollectionModal">
-        <ModalHeader>{modalTitle}</ModalHeader>
-        <ModalCloseButton />
-        <ModalBody pb={4}>
-          <AddCollectionForm {...formProps} />
-        </ModalBody>
-      </ModalContent>
+      <ModalHeader>{modalTitle}</ModalHeader>
+      <ModalCloseButton />
+      <AddCollectionForm {...formProps} />
     </Modal>
   );
 };